# server/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "e-Shopay",
        "description": "",
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/eshopay/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/cart/checkout": {
            "post": {
                "tags": [
                    "checkout"
                ],
                "summary": "Checkout single product from cart",
                "description": "",
                "operationId": "checkout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Checkout cart",
                        "required": true
                        
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            } 
        },
        "/cart/checkout/multiple": {
            "post": {
                "tags": [
                    "checkout"
                ],
                "summary": "Checkout multiple product from cart",
                "description": "",
                "operationId": "checkout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Checkout cart",
                        "required": true
                        
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            } 
        },
...
    }
}